+++ packages/flutter/lib/src/rendering/binding.dart
@@ -2,7 +2,6 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'dart:developer';
 import 'dart:ui' as ui show SemanticsUpdate;
 
 import 'package:flutter/foundation.dart';
@@ -507,13 +506,13 @@ mixin RendererBinding on BindingBase, ServicesBinding, SchedulerBinding, Gesture
     await super.performReassemble();
     if (BindingBase.debugReassembleConfig?.widgetName == null) {
       if (!kReleaseMode) {
-        Timeline.startSync('Preparing Hot Reload (layout)');
+        FlutterTimeline.startSync('Preparing Hot Reload (layout)');
       }
       try {
         renderView.reassemble();
       } finally {
         if (!kReleaseMode) {
-          Timeline.finishSync();
+          FlutterTimeline.finishSync();
         }
       }
     }
